+++ src/GslCore/Ape.fs
@@ -8,6 +8,12 @@ open Amyris.Bio.utils
 open utils
 open Amyris.Dna
 open Genbank
+open pragmaTypes
+        
+let topologyToString : Topology -> string =
+    function
+    | Linear -> "linear"
+    | Circular -> "circular" 
         
 /// Emit APE (genbank) format
 ///  outDir : string   tag: string  prefix for files  assemblies : List of AssemblyOut
@@ -22,7 +28,8 @@ let dumpAPE (outDir:string) (tag:string) (assemblies : DnaAssembly list) =
         let locusName = sprintf "%s_ape_output" tag
         let totLength = a.dnaParts |> List.map (fun p -> p.dna.Length) |> Seq.sum
         let now = DateTime.Now
-        sprintf "LOCUS                 %15s%5d bp ds-DNA   linear       %2d-%s-%d
+        let topology = a.topology |> topologyToString
+        sprintf "LOCUS                 %15s%5d bp ds-DNA   %s       %2d-%s-%d
 DEFINITION  .
 ACCESSION
 VERSION
@@ -31,7 +38,7 @@ SOURCE      .
 COMMENT
 COMMENT     ApEinfo:methylated:1
 FEATURES             Location/Qualifiers
-"           locusName totLength now.Day (mon.[now.Month-1]) now.Year |> w
+"           locusName totLength topology now.Day (mon.[now.Month-1]) now.Year |> w
         for p in a.dnaParts do
             let colorFwd = match p.sliceType with | REGULAR -> "#0000FF" | LINKER -> "#FF0000"
                                                   | MARKER -> "yellow" | INLINEST -> "green" | FUSIONST -> "red"