+++ src/GslCore/CommonTypes.fs
@@ -15,7 +15,6 @@ type SequenceLibrary = Map<string, Dna>
 /// Instructions gleaned from command line
 type ParsedOptions =
    {quiet: bool;
-    refStrain: string;
     libDir: string;
     iter: bool;
     onlyPhase1: bool;
@@ -210,16 +209,17 @@ let recalcOffset (pieces: DNASlice list) =
         {p with destFr = o; destTo = o+(p.dna.Length-1)*1<ZeroOffset> } )
 
 type DnaAssembly =
-   {id: int option;
-    dnaParts: DNASlice list;
-    name: string;
-    uri: Uri option;
-    linkerHint: string;
-    pragmas: PragmaCollection;
-    designParams: DesignParams;
-    docStrings: string list;
-    materializedFrom: Assembly}
-    with
+   { id : int option
+     dnaParts : DNASlice list
+     name : string
+     uri : Uri option
+     linkerHint : string
+     pragmas : PragmaCollection
+     designParams : DesignParams
+     docStrings : string list
+     materializedFrom : Assembly
+     topology : Topology }
+with
     member x.Sequence() =
         x.dnaParts
         |> Seq.map (fun p -> p.dna)